Pushing the store accessor to the square does nothing
- Time to investigate storing the board in the global store and reading from that
